pyproject.toml
[project]
name = "polymarket-mm-bot"
version = "2.0.0"
description = "Polymarket Market Maker Bot — triple revenue: spread + liquidity rewards + builder rewards"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "httpx>=0.27",
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "python-dotenv>=1.0",
    "py-clob-client>=0.18",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
    "mypy>=1.10",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
